--- students/MicahBraun/Lesson 06/Tests/.pytest_cache/v/cache/lastfailed
@@ -1,4 +1,0 @@
-{
-  "test_menu.py": true,
-  "test_menu.py::test_menu": true
-}
