--- students/MicahBraun/Lesson 06/Tests/.pytest_cache/v/cache/nodeids
@@ -1,3 +1,0 @@
-[
-  "test_menu.py::test_menu"
-]
