--- students/MicahBraun/Lesson 06/Tests/test_menu.py
@@ -1,14 +1,0 @@
-# ------------------------------------------------------------------------
-# NAME: MICAH BRAUN
-# PROJECT: 
-# PURPOSE: 
-# DATE: 
-#
-# DESCRIPTION:
-# ------------------------------------------------------------------------
-import magicmock
-from Tasks import mailroom_3_UPDATE
-
-
-def test_menu():
-    assert mailroom_3_UPDATE.menu() is function()
