+++ students/MicahBraun/Lesson 06/mailroom_refactor.py
@@ -1,0 +1,334 @@
+# ----------------------------------------------------------------------------------------------------------------------
+# NAME: MICAH BRAUN
+# PROJECT: mailroom_refactor.py - Lesson 06 Assignment
+# PURPOSE: TDD - Unit testing
+# DATE: 07/10/2018
+# EDITS MADE:
+#
+# 07/11/2018 -- write
+#
+# 07/10/2018 -- refactored thankyou() broken up in to 3 functions (add_donors, add_donoation, thankyou)
+# this was done to make unit testing more manageable and to make the code easier to read.
+#
+# 06/29/2018 -- Try, Except blocks added for exception handling, .get() removed from menu(),
+# 'pair' var updated to handle error on quitting before assigning donation value to new name entry in thankyou()
+#
+#  DESCRIPTION: Mailroom.py takes user input and expands list of donor information to be displayed/manipulated
+#  back to the user.
+# ----------------------------------------------------------------------------------------------------------------------
+import time     # time to supply today's date for letters
+import os       # for default working directory
+import locale   # for currency formatting
+import sys      # for error handling
+#  --------------------------------------------------- DATA ------------------------------------------------------------
+
+locale.setlocale(locale.LC_ALL, '')
+today = time.strftime("%m/%d/%Y")
+
+
+spacing = '- ' * 56  # formatting for DONOR header
+donor_dict = {'Columns': ['NAMES', 'DONATION AMOUNT', 'NUMBER OF GIFTS', 'AVG. GIFTS'],
+              'Rudolph Soares': ['Rudolph Soares', 1335, 3, 445],  # Donor info
+              'Josef Mistretta': ['Josef Mistretta', 7500, 4, 1875],
+              'Joye Agostini': ['Joye Agostini', 600, 2, 300],
+              'Rachelle Levan': ['Rachelle Levan', 12100, 5, 2420],
+              'Vena Ussery': ['Vena Ussery', 4000, 8, 500],
+              'Efrain Lager': ['Efrain Lager', 795, 9, 88.34],
+              'Mee Heine': ['Mee Heine', 4600, 4, 1150],
+              'Tanya Essex': ['Tanya Essex', 75000, 2, 37500]}
+
+
+#  ------------------------------------------------- PROCESSING --------------------------------------------------------
+def menu():
+    """Display menu of options to toggle between to perform separate actions"""
+    header = "-" * 110  # formatting for header
+    title = ' ' * 48 + "MAIL ROOM MENU"  # title display
+    directions = "Select one of the three options:"  # directions to user for menu
+    optA = "[A] Send a Thank you"  # option(s) for menu
+    optB = "[B] Create a Report"
+    optC = "[C] Create Letters for All Donors"
+    optD = "[D] Quit"
+    whole_menu_header = header + '\n' + title + '\n' + header  # header strung together as one unit
+
+    while True:
+        print(whole_menu_header)  # display header
+        print(directions.center(110, ' '), '\n')  # center directions with header display
+        print(optA + optB.center(37, ' ') + optC.center(36, ' ') + optD.center(25, ' ') + '\n')
+
+        switch_function = {
+            'A': add_donors,  # dictionary holding menu options
+            'B': createreport,
+            'C': writeletters,
+            'D': quit_program,
+        }
+        try:
+            usr_sel = (input("Menu Selection: ")).upper()  # user input for menu
+            switch_function[usr_sel]()  # get user entry from dict
+
+        except KeyError:
+            print("Invalid Entry! Only enter A, B, C, or D.")  # display error on any other entry
+        continue  # continue back to start of menu
+
+
+def add_donors():
+    """Add donor names to list"""
+    print("Leaving menu...\n")
+    while True:
+        print("Enter the name of the person you are writing to (or enter 'list' to see a list of names or Q to quit) ")
+        fname_prompt = input("First Name: ").strip().capitalize()  # first name variable (strip any spaces)
+        if fname_prompt.upper() == "Q":  # if Q, then quit to menu()
+            menu()
+        elif fname_prompt.lower() == "list":
+            displaylist()
+        else:
+            lname_prompt = input("Last Name: ").strip().capitalize()  # last name variable
+            if lname_prompt.upper() == "Q":  # if Q, back to menu
+                menu()
+            elif lname_prompt.lower() == "list":  # if list, then list display
+                displaylist()
+            else:
+                key = fname_prompt + " " + lname_prompt
+                if key in donor_dict.keys():
+                    existing_donor = input("That value is already in the list! Do you want to "
+                                           "proceed with that selection? (Y/N): ")
+                    if existing_donor.upper() == "Y":  # if user proceeds, print display
+                        add_donation(key)
+                else:
+                    add_name = str(input("That name is not in the Donor list. Do you want to add it to the list? (Y/N) "))
+                    if add_name.upper() == "Y":  # if input response == "Y", proceed
+                        pair = {key: [key, 0, 0, 0]}
+                        donor_dict.update(pair)
+                        displaylist()
+                        # num_donations = 1
+                        add_donation(key)
+
+
+def displaylist():
+    """Function returns current dictionary contents (names only)"""
+    cut_off = int(len(list(donor_dict.values())) / 2)  # where list breaks for next line
+    if int(len(list(donor_dict.values()))) - 1 % 2 != 0:  # even/odd check
+        for i in range(0, int(len(list(donor_dict.values())) - int(len(list(donor_dict.values())) - 2) / 2)):
+            if i == 0:  # header check
+                print(spacing)
+                print('{:>44s}'.format(str(list(donor_dict.values())[i][0])))  # print Donor Name header
+                print(spacing)
+            elif cut_off + i >= len(list(donor_dict.values())):  # if cut_off + i exceeds len(list)
+                print('{:>30s}'.format(list(donor_dict.values())[i][0]))  # only print i (not cut_off + i)
+                continue
+            else:
+                print('{:>30s}'.format(list(donor_dict.values())[i][0]),
+                      '{:>35s}'.format(list(donor_dict.values())[cut_off + i][0]))
+    else:
+        for i in range(0, int(len(list(donor_dict.values())) / 2)):
+            if i == 0:  # header check
+                print(spacing)
+                print('{:>44s}'.format(str(list(donor_dict.values())[i][0])))  # print Donor Name header
+                print(spacing)
+            else:
+                print('{:>30s}'.format(list(donor_dict.values())[i][0]),
+                      '{:>35s}'.format(list(donor_dict.values())[cut_off + i][0]))
+    print()
+
+
+def add_donation(name):
+    """Add donation amounts"""
+    donation_amt_str = input("Enter in the donation amount from Donor {0}: $".format(name))
+    if donation_amt_str.lower() == 'q':
+        menu()
+    try:
+        donation_amt = int(donation_amt_str)
+    except ValueError:
+        print("Error: invalid entry.\n")
+    else:
+        print('{0} has donated ${1:,.2f}'.format(name, donation_amt))  # display name and donation amt
+        int(donation_amt)
+        thankyou(name, donation_amt)
+
+
+def thankyou(donor, donation):
+    """Function enables user to add to dictionary data (adding new donors and donation amounts as well as
+    number of donations and average donations for new keys, and also updates existing keys."""
+    get_index = 0  # index variable
+    for item in range(0, len(donor_dict)):  # set index var to current name
+        if list(donor_dict.values())[item][0] == donor:
+            get_index = item
+            break
+    firstname = list(donor_dict.values())[get_index][0].split(' ', ).pop(0)  # separate first name
+    current_donations = int(list(donor_dict.values())[get_index][1])  # current amt/ var from dict{[val]}
+    sum_donations = current_donations + donation  # sum of all donations (current + new)
+    list(donor_dict.values())[get_index][1] = float(sum_donations)  # update sum
+    num_donations = int(list(donor_dict.values())[get_index][2]) + 1  # num donations = self + 1
+    list(donor_dict.values())[get_index][2] = num_donations  # update num of donations
+    list(donor_dict.values())[get_index][3] = averagedonations(sum_donations, num_donations)  # update avg
+    email_display = str(input("Display Donor Email? (Y/N or Q): "))  # ask user if they want to to
+    if email_display.upper() == "Y":  # write email with name/amount
+        print(spacing)
+        print('Dear {0}, \n\nThank you for your continued support through your \
+contribution of ${1:,.2f} towards our Foundation\'s fundraising goal.\n\nBest wishes,\n\
+Foundation Board of Directors\n'.format(firstname, sum_donations))
+        print(spacing)
+    elif email_display.upper() == 'Q':
+        menu()
+    else:
+        pass
+
+
+def createreport():
+    """Function creates organized report of donor information based on current dictionary contents.
+    Information is sorted in descending order based on total donation amounts."""
+    print("Leaving menu...")
+    while True:  # continual loop unless user terminates
+        proceed = str(input("Generate Donor report? (Y/N): "))  # solicit user response to proceed
+        if proceed.upper() == "Y":  # if user indicates yes
+            input("Generating Donor report... [Press Enter]")  # on Enter press, proceed
+            new_list = list(donor_dict.values())[1:]  # blank list to hold only values after headers
+            new_list.sort(key=lambda sort_on: sort_on[1], reverse=True)  # sort list based on donation totals
+            lst_heading = list(donor_dict.values())[0:1]  # container for only header
+            final_lst = lst_heading + new_list  # join sorted list back with header
+            print(spacing)
+            print('{:>15s}'.format(str(final_lst[0][0])), end='')  # format header for display
+            print('         |{:>20s}'.format(str(final_lst[0][1])), end='')
+            print('   |{:>19s}'.format(str(final_lst[0][2])), end='')
+            print('    |{:>15s}'.format(str(final_lst[0][3])))
+            print(spacing)
+            for i in range(1, int(len(final_lst))):  # format list contents
+                print('{:>20s}'.format(str(final_lst[i][0])), end='')
+                print('            ${:>10,.2f}{:>17}'.format(float(final_lst[i][1]), int(final_lst[i][2])), end='')
+                print('                 ${:>10,.2f}'.format(float(final_lst[i][3])))
+        else:
+            back_to_menu = str(input("Do you want to quit back to the main menu? (Y/N): "))  # escape sequence
+            if back_to_menu.upper() == "Y":
+                input("Quitting... [Press Enter]")
+                menu()
+            else:
+                continue  # if no esc, then continue loop
+
+
+def get_directory():
+    """Obtain or assign default directory for writeletters()"""
+    default = os.getcwd()  # get and assign current directory to variable
+    default = default + "\\"  # format default (will work with Windows)
+    try:
+        working_directory = str(input('\nEnter the file path where you want to write the letters - '
+                                      'don\'t forget to use ''two \n\'\\\\\' as file separators to make sure'
+                                      ' it is compatible if using Windows - (e.g. \'C:\\\\\': '))
+    except Exception as e:              # exception handler - prints line of error, and type
+        print('Error on line {}'.format(sys.exc_info()[-1].tb_lineno), type(e).__name__, e)
+        # working directory variable assignment by user - if blank, use default
+
+    else:
+        if working_directory != "":  # check for input on working_directory
+            return working_directory
+        else:
+            working_directory = default
+            return working_directory
+
+
+def writeletters():
+    """Function writes donor letter to file in user-directed file path"""
+    print("Leaving menu...")
+
+    names_list = []  # empty list to hold list of donor_dict names w/o whitespace (for filenames)
+    first_names = []  # first names separated out to be used for letters
+    while True:
+        msg = input("Proceed to write letters to all donors in donor list? (Y/N) ")  # if user selects Y, proceed
+        if msg.upper() == "Y":  # check input
+            cur_dir = get_directory()
+            for x in range(int(len(donor_dict))):  # for items in span donor_dict
+                firstname = list(donor_dict.values())[x][0]  # firstname = donor_dict at index[item], column[0]
+                firstname = firstname.split(' ', ).pop(0)  # format to split on whitespace
+                first_names.append(firstname)  # append value to empty list
+
+            for i in range(int(len(list(donor_dict.values())))):  # for items in the span of donor_dict
+                elem = list(donor_dict.values())[i][0]  # elem = donor_dict at index[i], column[0]
+                elem = elem.replace(" ", "")  # format name w/o whitespace
+                names_list.append(elem)  # append to name_list
+
+            for names in range(1, int(len(names_list))):  # omit header, for each name in names_list[]
+                try:
+                    with open(cur_dir + str(names_list[names]) + '.txt',
+                              'a') as filename:  # create files/file-names
+                        filename.write(today)  # write current date (mm/dd/yy)
+                        filename.write('\n\nDear {0}, \n\nThank you for your continued support through your contribution '
+                                       'of {1} towards our Foundation\'s fundraising goal.\n\nBest wishes,\n'
+                                       'Foundation Board of Directors'
+                                       '\n'.format(str(first_names[names]),
+                                                   locale.currency(list(donor_dict.values())[names][1], grouping=True)))
+                except Exception as e:
+                    print('Error on line {}'.format(sys.exc_info()[-1].tb_lineno), type(e).__name__, e)
+            print("Files being written to " + cur_dir + "...")
+            print("Done!")
+
+        elif msg.upper() == "N":  # if "N" for msg = input()
+            input("Exiting back to menu [Press ENTER]")  # leaving this function
+            break  # back to menu()
+        else:
+            print("That's an unrecognized value! Only Y or N are accepted.")
+            continue  # warning message: back to beginning of function
+
+
+def quit_program():
+    """Function quits all processes"""
+    input("Exiting Program...[Press ENTER]")    # I use PyCharm and exit() w/o import worked fine but
+    from sys import exit                        # iPython was not running exit() without importing it from sys
+    exit()                                      # so to avoid issues, I just kept this here
+
+
+def averagedonations(donations, num):
+    """Returns average of donor's gifts"""
+    sumgifts = donations
+    numgifts = num
+
+    average_donation = sumgifts / numgifts
+    return average_donation
+
+
+'''
+#   NOTE: I could not get these 2 functions to work together... I tried breaking up the displaylist() function
+          into a processing function and a display function to better separate things for test-ability but
+          I kept running into typeErrors, inaccessible values within if-blocks, and the final attempt yielded
+          just blank returns. Not sure what to do with this section
+          
+def return_list():
+    """
+    Container for list information
+    """
+    cut_off = int(len(list(donor_dict.values())) / 2)  # where list breaks for next line
+    global combine
+    combine = []
+    if int(len(list(donor_dict.values()))) - 1 % 2 != 0:  # even/odd check
+        for i in range(0, int(len(list(donor_dict.values())) - int(len(list(donor_dict.values())) - 2) / 2)):
+            if i == 0:  # header check
+                list_header = spacing + '{:>44s}'.format(str(list(donor_dict.values())[i][0])) + spacing
+                combine.append(list_header)
+            elif cut_off + i >= len(list(donor_dict.values())):  # if cut_off + i exceeds len(list)
+                uneven_left_col_handler = '{:>30s}'.format(list(donor_dict.values())[i][0])  # i (not cut_off + i)
+                combine.append(uneven_left_col_handler)
+                continue
+            else:
+                uneven_right_col_handler = '{:>30s}'.format(list(donor_dict.values())[i][0]), \
+                                           '{:>35s}'.format(list(donor_dict.values())[cut_off + i][0])
+                combine.append(uneven_right_col_handler)
+        return combine
+
+    else:
+        for i in range(0, int(len(list(donor_dict.values())) / 2)):
+            if i == 0:  # header check
+                list_header = spacing + '{:>44s}'.format(str(list(donor_dict.values())[i][0])) + spacing
+                combine.append(list_header)
+            else:
+                col_handler = '{:>30s}'.format(list(donor_dict.values())[i][0]), \
+                                           '{:>35s}'.format(list(donor_dict.values())[cut_off + i][0])
+                combine.append(col_handler)
+    return combine
+
+
+def print_list():
+    """Function returns current dictionary contents (names only)"""
+    list_display = return_list()
+    print(list_display)
+'''
+
+#  -------------------------------------------------- DISPLAY ----------------------------------------------------------
+if __name__ == '__main__':
+    menu()  # Program starts here
